requirements.txt
Flask==2.3.3
alpaca-trade-api==3.1.0
python-dotenv==1.0.1
gunicorn==21.2.0
